Adding 'core' to setting.py and writing TDD test cases
We write test case to verify that a django user can be created
by passing 'email' and 'password'.
We then run the command -
  docker-compose run app sh -c "python manage.py test && flake8"
to verify whether the test works or not. And as expected
the test fails (because by default django auth user's create_user()
method expects a username and we are passing an email address).